Make all elements value types instead of classes
Got inspired by these WWDC 2015 videos:

- Protocol-Oriented Programming in Swift:
	https://developer.apple.com/videos/play/wwdc2015/408/

- Building Better Apps with Value Types in Swift:
	https://developer.apple.com/videos/play/wwdc2015/414/

We already had value types for our attributes, but reimplementing elements as values took some rethinking, particularly since the parser broke when just changing class to struct. To make it work, instead of building up our element tree by adding children as the elements begin parsing, we essentially have to build it up backwards. That is, whenever endElement: is called, we chop said element off the stack, then we take the next top element off the stack and try to add our current element to it. Then we re-add that parent to the stack.

It makes sense now that I wrapped my head around it, but there was definitely some headbanging going on initially. I do believe this will result in a better architecture for this framework.